move test utils and update expected metric size
