+++ x-pack/plugins/files/server/routes/file_kind/integration_tests/file_kind_http.test.ts
@@ -6,7 +6,7 @@
  */
 
 import type { UpdatableFileMetadata } from '../../../../common/types';
-import { setupIntegrationEnvironment, TestEnvironmentUtils } from '../../test_utils';
+import { setupIntegrationEnvironment, TestEnvironmentUtils } from '../../../test_utils';
 
 describe('File kind HTTP API', () => {
   let fileKind: string;
